backend/requirements.txt
# Core FastAPI
fastapi
uvicorn[standard]

# Database
sqlalchemy
python-dotenv
psycopg2-binary

# Pydantic email validation
email-validator

# AI Features (optional)
openai

# Essential Security (Manual Entry Apps)
python-jose[cryptography]  # JWT tokens
passlib[bcrypt]==1.7.4    # Password hashing
bcrypt==4.1.3             # Compatible bcrypt version
python-multipart          # Form handling

# Optional: Enhanced Security (add later if needed)
# slowapi                 # Rate limiting
# cryptography           # Additional encryption
# fastapi-mail           # Email verification 